Refactored Bid to a seperate context
