+++ frontend/src/Pages/AuctionPage/AuctionPage.tsx
@@ -6,7 +6,8 @@ import {
   StyledSearchFieldWrapper,
 } from "./StyledAuctionPage";
 import AuctionCard from "../../Components/AuctionCard/AuctionCard";
-import { useAuction, Bid } from "../../Contexts/AuctionContext";
+import { useAuction } from "../../Contexts/AuctionContext";
+import { Bid } from "../../Contexts/BidContext";
 import { useEffect } from "react";
 import SearchForm from "../../Components/Search/SearchForm/SearchForm";
 import { useSearch } from "../../Contexts/SearchContext";
